requirements.txt
pydub
openai-whisper
torch